Moving login status from tripsTable.jsx to TripContext.jsx, making it easier to access by exporting it for all children components
